refactor: useWalkAvailable 오타 수정
