+++ frontend/src/hooks/useWalkAvailableDog.ts
